gravity, collision, thrust
added gravity similar to the gravity on the moon. made collisions more reliable. added a thrust script that can be edited in the unity inspector. apply thrust using the space bar. thrust must be able to move the object it is applied to based on the mass of the object and the gravity of the environment.