Cube rotation thrusting
rotations added. a new cube has been added to apply the rotations and thrust too. the ball is unstable because it rolls. the cube now lifts off, rotates, and moves based on thrust and the direction it is facing from the rotations. the cube is followed by the camera.

fuel tank needs added
rotation locks to prevent too much rotation needs added